requirements.txt
numpy
torch
Pillow
torchvision
tensorboard
scikit-learn
matplotlib
requests
tqdm